@@ -8,41 +8,33 @@ namespace MinimalApiBuilder.Generator.CodeGeneration;
 internal abstract class SourceBuilder
 {
     private const int IndentSize = 4;
-    private readonly StringBuilder _builder = new(@"// <auto-generated>
-// This is a MinimalApiBuilder source generated file.
-// </auto-generated>
-
-#nullable enable
-
-");
+    private readonly StringBuilder _builder = new(Sources.Header);
     private int _indent;
-    protected static readonly string s_generatedCodeAttribute =
-        $@"[global::System.CodeDom.Compiler.GeneratedCodeAttribute(""{typeof(SourceBuilder).Assembly.GetName().Name}"", ""{typeof(SourceBuilder).Assembly.GetName().Version}"")]";
+
+    protected SourceBuilder() : this(default)
+    { }
 
     protected SourceBuilder(GeneratorOptions options)
     {
         Options = options;
+        _builder.AppendLine();
     }
 
     protected GeneratorOptions Options { get; set; }
 
     protected ICollection<Diagnostic> Diagnostics { get; } = new List<Diagnostic>();
 
+    public sealed override string ToString() => _builder.ToString();
+
     public abstract void AddSource(SourceProductionContext context);
 
-    public void ReportDiagnostics(SourceProductionContext context)
+    protected void AddSource(SourceProductionContext context, string hintName)
     {
+        context.AddSource($"{hintName}.g.cs", ToString());
         foreach (Diagnostic diagnostic in Diagnostics)
         {
             context.ReportDiagnostic(diagnostic);
         }
-    }
-
-    public sealed override string ToString() => _builder.ToString();
-
-    protected void MarkAsGenerated()
-    {
-        AppendLine(s_generatedCodeAttribute);
     }
 
     protected IDisposable OpenBlock(string value)
@@ -83,6 +75,11 @@ internal abstract class SourceBuilder
         }
 
         return OpenBlockExtra(afterClose, values[values.Length - 1]);
+    }
+
+    protected void AppendLine()
+    {
+        _builder.AppendLine();
     }
 
     protected void AppendLine(string value)
