@@ -1,16 +1,17 @@
 ﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
 using MinimalApiBuilder;
 
 namespace Fixture.TestApi.Features.Validation;
 
 internal partial class CombinedValidationEndpoint : MinimalApiBuilderEndpoint
 {
-    private static IResult Handle(
+    public static IResult Handle(
         [AsParameters] SyncValidationParameters parameters,
         AsyncValidationRequest request,
-        [FromServices] CombinedValidationEndpoint endpoint)
+        Serilog.ILogger logger)
     {
+        logger.Information("Parameters: {Parameters}", parameters);
+        logger.Information("Request: {Request}", request);
         return TypedResults.Ok();
     }
 }
