@@ -52,25 +52,84 @@ public partial class E : MinimalApiBuilderEndpoint
     [Test]
     public Task Endpoint_Parameter_At_Different_Locations([Values(0, 1, 2, 3)] int location)
     {
-        List<string> parameters = new()
-        {
-            "int a",
-            "R r",
-            "int b"
-        };
+        List<string> parameters = ["int a", "R r", "int b"];
         parameters.Insert(location, "E e");
-
         string parametersString = string.Join(", ", parameters);
 
         // lang=cs
         string source = $$"""
 public class R {
     public int Value { get; set; }
+    public static ValueTask<R> BindAsync(HttpContext context)
+    {
+        return ValueTask.FromResult<R>(new R());
+    }
 }
 
 public partial class E : MinimalApiBuilderEndpoint
 {
     public static int Handle({{parametersString}}) => 0;
+}
+
+public class RValidator : AbstractValidator<R>
+{
+    public RValidator()
+    {
+        RuleFor(static x => x.Value).GreaterThan(0);
+    }
+}
+""";
+
+        return VerifyGeneratorAsync(source);
+    }
+
+    [Test]
+    public Task Without_Endpoint_Parameter()
+    {
+        // lang=cs
+        const string source = """
+public class R {
+    public int Value { get; set; }
+}
+
+public partial class E : MinimalApiBuilderEndpoint
+{
+    public static int Handle(int a, int b, R r) => a;
+}
+
+public class RValidator : AbstractValidator<R>
+{
+    public RValidator()
+    {
+        RuleFor(static x => x.Value).GreaterThan(0);
+    }
+}
+""";
+
+        // lang=cs
+        const string mapActions = """
+app.MapDelete("/test/{a:int}/{b:int}", E.Handle);
+""";
+
+        return VerifyGeneratorAsync(source, mapActions);
+    }
+
+    [Test]
+    public Task Without_Endpoint_Parameter_And_Custom_Binding()
+    {
+        // lang=cs
+        const string source = """
+public class R {
+    public int Value { get; set; }
+    public static ValueTask<R> BindAsync(HttpContext context)
+    {
+        return ValueTask.FromResult<R>(new R());
+    }
+}
+
+public partial class E : MinimalApiBuilderEndpoint
+{
+    public static int Handle(int a, int b, R r) => a;
 }
 
 public class RValidator : AbstractValidator<R>
